[invoker] Fix delay UINT_MAX check
Found when compiling for x86_64 under clang:

src/invoker/invoker.c:638:39: warning: result of comparison of constant 18446744073709551615 with expression of type 'unsigned int' is always false [-Wtautological-constant-out-of-range-compare]
  638 |         if ((errno == ERANGE && delay == ULONG_MAX)
      |                                 ~~~~~ ^  ~~~~~~~~~

+++ src/invoker/invoker.c
@@ -636,7 +636,7 @@ static unsigned int get_delay(char *delay_arg, char *param_name,
         delay = strtoul(delay_arg, NULL, 10);
 
         // Check for various possible errors
-        if ((errno == ERANGE && delay == ULONG_MAX)
+        if ((errno == ERANGE && delay == UINT_MAX)
             || delay < min_value
             || delay > max_value)
         {
